refactor: extract types to new files
