+++ lib/config/presets/types.ts
@@ -13,3 +13,11 @@ export type PresetConfig = {
 export interface PresetApi {
   getPreset(config: PresetConfig): Promise<Preset> | Preset;
 }
+
+export interface ParsedPreset {
+  presetSource: string;
+  packageName: string;
+  presetPath?: string;
+  presetName: string;
+  params?: string[];
+}
