+++ lib/config/validation-helpers/managers.ts
@@ -1,10 +1,6 @@
 import { getManagerList } from '../../manager';
-import type { PackageRule, ValidationMessage } from '../types';
-
-export interface CheckManagerArgs {
-  resolvedRule: PackageRule;
-  currentPath: string;
-}
+import type { ValidationMessage } from '../types';
+import type { CheckManagerArgs } from './types';
 
 /**
  * Only if type condition or context condition violated then errors array will be mutated to store metadata
