+++ lib/config/validation-helpers/types.ts
@@ -1,0 +1,6 @@
+import { PackageRule } from '../types';
+
+export interface CheckManagerArgs {
+  resolvedRule: PackageRule;
+  currentPath: string;
+}
