+++ lib/datasource/cdnjs/types.ts
@@ -1,0 +1,13 @@
+interface CdnjsAsset {
+  version: string;
+  files: string[];
+  sri?: Record<string, string>;
+}
+export interface CdnjsResponse {
+  homepage?: string;
+  repository?: {
+    type: 'git' | unknown;
+    url?: string;
+  };
+  assets?: CdnjsAsset[];
+}
