+++ lib/datasource/dart/types.ts
@@ -1,0 +1,9 @@
+export interface DartResult {
+  versions?: {
+    version: string;
+    published?: string;
+  }[];
+  latest?: {
+    pubspec?: { homepage?: string; repository?: string };
+  };
+}
