+++ lib/datasource/docker/index.ts
@@ -14,7 +14,7 @@ import type { OutgoingHttpHeaders } from '../../util/http/types';
 import { ensureTrailingSlash, trimTrailingSlash } from '../../util/url';
 import * as dockerVersioning from '../../versioning/docker';
 import type { GetReleasesConfig, ReleaseResult } from '../types';
-import { Image, ImageList, MediaType } from './types';
+import { Image, ImageList, MediaType, RegistryRepository } from './types';
 
 // TODO: add got typings when available (#9646)
 // TODO: replace www-authenticate with https://www.npmjs.com/package/auth-header (#9645)
@@ -55,11 +55,6 @@ export const defaultConfig = {
 const http = new Http(id);
 
 const ecrRegex = /\d+\.dkr\.ecr\.([-a-z0-9]+)\.amazonaws\.com/;
-
-export interface RegistryRepository {
-  registry: string;
-  repository: string;
-}
 
 export function getRegistryRepository(
   lookupName: string,
