+++ lib/datasource/git-refs/index.ts
@@ -3,6 +3,7 @@ import * as packageCache from '../../util/cache/package';
 import { getRemoteUrlWithToken } from '../../util/git/url';
 import * as semver from '../../versioning/semver';
 import type { DigestConfig, GetReleasesConfig, ReleaseResult } from '../types';
+import type { RawRefs } from './types';
 
 export const id = 'git-refs';
 export const customRegistrySupport = false;
@@ -11,12 +12,6 @@ const cacheMinutes = 10;
 
 // git will prompt for known hosts or passwords, unless we activate BatchMode
 process.env.GIT_SSH_COMMAND = 'ssh -o BatchMode=yes';
-
-export interface RawRefs {
-  type: string;
-  value: string;
-  hash: string;
-}
 
 export async function getRawRefs({
   lookupName,
