+++ lib/datasource/git-refs/types.ts
@@ -1,0 +1,5 @@
+export interface RawRefs {
+  type: string;
+  value: string;
+  hash: string;
+}
