+++ lib/datasource/github-tags/types.ts
@@ -1,0 +1,7 @@
+export interface TagResponse {
+  object: {
+    type: string;
+    url: string;
+    sha: string;
+  };
+}
