+++ lib/datasource/go/types.ts
@@ -1,0 +1,5 @@
+export interface DataSource {
+  datasource: string;
+  registryUrl?: string;
+  lookupName: string;
+}
