+++ lib/datasource/hex/index.ts
@@ -3,6 +3,7 @@ import { ExternalHostError } from '../../types/errors/external-host-error';
 import { Http } from '../../util/http';
 import * as hexVersioning from '../../versioning/hex';
 import type { GetReleasesConfig, ReleaseResult } from '../types';
+import type { HexRelease } from './types';
 
 export const id = 'hex';
 export const defaultRegistryUrls = ['https://hex.pm/'];
@@ -10,15 +11,6 @@ export const customRegistrySupport = false;
 export const defaultVersioning = hexVersioning.id;
 
 const http = new Http(id);
-
-interface HexRelease {
-  html_url: string;
-  meta?: { links?: Record<string, string> };
-  releases?: {
-    version: string;
-    inserted_at?: string;
-  }[];
-}
 
 export async function getReleases({
   lookupName,
