+++ lib/datasource/hex/types.ts
@@ -1,0 +1,8 @@
+export interface HexRelease {
+  html_url: string;
+  meta?: { links?: Record<string, string> };
+  releases?: {
+    version: string;
+    inserted_at?: string;
+  }[];
+}
