+++ lib/datasource/orb/types.ts
@@ -1,0 +1,7 @@
+export interface OrbRelease {
+  homeUrl?: string;
+  versions: {
+    version: string;
+    createdAt?: string;
+  }[];
+}
