+++ lib/datasource/terraform-provider/index.ts
@@ -5,6 +5,10 @@ import { Http } from '../../util/http';
 import * as hashicorpVersioning from '../../versioning/hashicorp';
 import { getTerraformServiceDiscoveryResult } from '../terraform-module';
 import type { GetReleasesConfig, ReleaseResult } from '../types';
+import type {
+  TerraformProvider,
+  TerraformProviderReleaseBackend,
+} from './types';
 
 export const id = 'terraform-provider';
 export const customRegistrySupport = true;
@@ -16,27 +20,6 @@ export const defaultVersioning = hashicorpVersioning.id;
 export const registryStrategy = 'hunt';
 
 const http = new Http(id);
-
-interface TerraformProvider {
-  namespace: string;
-  name: string;
-  provider: string;
-  source?: string;
-  versions: string[];
-  version: string;
-  published_at: string;
-}
-
-interface TerraformProviderReleaseBackend {
-  [key: string]: {
-    name: string;
-    versions: VersionsReleaseBackend;
-  };
-}
-
-interface VersionsReleaseBackend {
-  [key: string]: Record<string, any>;
-}
 
 async function queryRegistry(
   lookupName: string,
